net: fec: avb: make the ptp counter oper in 32 bits
commit 25b1a6b62e14b09e0752ef1308c0c7a451792c43 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git

The AVB stack still expects ptp timestamps to be on 32 bits, so implement
the workaround to make the clock work with that (instead of 31 bits).

Modify the FEC PTP clock driver to update the HW ENET counter when
processing clock_adj{time,freq} syscalls, instead of simply maintaining
ratio and offsets in software.
This will allow network timestamps to match the PTP time, and to trigger
DMA events based on PTP timestamps without the need for any timescale
translation.

Note: Since 3.14.28 Freescale BSP does update the HW counter. However
as a WA of a 1588 HW issue, the counter is limited to 31 bits and the AVB stack
expects timestamps to be on 32 bits. So for now we still need this patch.

Make FEC PTP clock adjust to work around hardware issue:
. Compute period for twice the requested correction
. Use only even period values

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>