net: fec: avb: add avb transmit path
commit 44cf0aceabe6f21f32029881d49fd9326574ee09 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git

- Integrate FEC driver transmit with AVB transmit layer.
When AVB is enabled all packets transmitted through the regular
network stack are put in a best effort queue and transmitted based on
802.1Qat QoS algorithm.

- Modified tx avb interface (to workaround tx errata):
Change the avb transmit interface to workaround an ethernet tx errata
(where the hardware may stop transmit with packets pending, if it
previously saw a "not ready" descriptor). A new function is added
to trigger the actual transmit (for devices affected by errata).
Typically the qos scheduler will wait for the last packet, in a
scheduling period to be posted, before calling the new function.
This assures the hardware can only see "ready" descriptors once
it's started and until the next scheduling period.
For a device not affected by the errata, transmit is still triggered
for each packet (and calling the new funcion has no effect).

- Also, setting the set FTYPE flag for all frames (AVTP and best effort)
on AVB capable boards on fec_enet_start_xmit_avb

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>