arm64: KVM: support flushing device memory
commit 9b062acf5434161bb89afff798c86f2a7ef8afd5 from
https://github.com/nxp-imx/linux-imx

In the current implementation, trying to flush
memory not covered by the linear map (e.g. device memory) causes
a crash.
The QMAN region, even though it is a device region, it's cacheable.
However, as we map the qman region with the same attributes as
the guest, there is no need for cache flush. Use some reserved
bits (reserved for software use) from the descriptor to mark this
type of memory.
This allows dropping the special checks for qman
cacheable region when doing cache flushes.

Signed-off-by: Laurentiu Tudor <laurentiu.tudor@nxp.com>
[rebased on upstream changes in KVM memory management -
using generic page table API + hyp memory isolation]
Signed-off-by: Diana Craciun <diana.craciun@oss.nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>