MLK-13308-2 usb: phy: phy-mxs-usb: handle USB PHY event
commit  037e8dc563d503667ffecc321328cb412ab8f145 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

For mxs PHY, if there is a vbus but the bus is not enumerated,
force the dp/dm as SE0 from the consider side. If not, there
is possible USB wakeup due to unstable dp/dm, since there is
possible no pull on dp/dm, eg, there is a USB charger on the
port. Note, the vbus event is only occurred at device mode,
and sent by udc driver.

Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>