crypto: caam - add support for MOVEB command
commit  6275526fe7a9105c8b7c2c8611db1c8de54b27ca from
https://github.com/nxp-imx/linux-imx lf-6.1.y

CHAs of SEC work natively in BE mode. When moving
data to the alignment blocks, swapping is needed
for LE platforms. This is done by means of the MOVEB
command. This patch adds support
to DCL for this command.

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Signed-off-by: Radu Alexe <radu.alexe@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>