MLK-17537-1 drm/bridge: Correct the ADV7533 init
commit 4d11e91b88929aa027fb7117d8958b212539b253 from
https://github.com/nxp-imx/linux-imx

According to the Analog Devices configuration script, there are some
steps that need to be made when configuring the ADV for a specific mode.
Some of those steps were missing from driver, so this patch takes care
of this.
Also, in mode_fixup, the driver is trying to reconfigure the DSI lanes
from 4 to 3, when pixel clock is lower than 80MHz, which is not
necessary. the lanes property represents the maximum available lanes on
that devices and should not differ from a mode to another.
The DSI host is the one who should predict how many lanes it could use
to drive a display mode, so remove this from ADV driver.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
[Xiaolei: "adv7533_mode_set" This part of the
modification has been submitted by upstream,
xiaolei delete this part of the code.]
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>