drm/imx: sec-dsim_imx: a general way to compute PLL PMS
commit e5b89c4da1487e521fe7d108b1bdbff0c80ebe62 from
https://github.com/nxp-imx/linux-imx

A fixed PLL PMS setting for attached panel is obviously not
enough for any other mipi panel which needs a different PLL
output clock frequency, and besides, for the CEA-861 standard
display modes, the 'pll_pms' table also can not cover all the
modes requirements. So a general way is created to solve this
problem which can provide an optimum solution to output a PLL
bit clock to match the request frequency in a maximum degree
and also satisfy the input clock and intermediate clocks limit
according to the PLL specification.

(This is the DSIM controller driver change for PLL PMS compute.)

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>