sdk_dpaa: ceetm: fix recursive dependencies
commit  cddba24c8b0d604e6ea4f4cd60a744239d082090 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

Due to dependencies between the fsl_ceetm and fsl_dpa modules, remove
the module support for the ceetm driver and integrate it into the main
DPAA Ethernet driver.

The registration of the CEETM Qdisc was the only operation done at
module init. Pass the management of the Qdisc register and unregister
operations to the loading and unloading of the DPAA driver.

Signed-off-by: Camelia Groza <camelia.groza@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>