usb: chipidea: add controller resume support when controler is powered off
commit  0791d25578cb0e46fd93ae7a3c36ff7a424f3547 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

For some SoCs, the controler's power will be off during the system
suspend, and it needs some recovery operation to let the system
back to workable. We add this support in this patch. It needs to
split some small patches for upstream in future.

Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>