dt-bindings: clocks: imx8mp: Add IDs for SAI MCLKs
commit f7b6eed58c77f4030166f25effa47c3585d16b59 from
https://github.com/nxp-imx/linux-imx

These IDs will be used by the platform clock controller driver
to register the clocks even if they are not populated on
a specific board. And that's okay, because it will allow the
driver to remain unchanged, the only thing changing will be the
dtb between boards which will override the properties of each
specific SAI MCLK devicetree node as necessary, if the board
has populated the any of the SAIx_MCLK pins with an input clock.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Reviewed-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>