clk: imx8mp: Register the SAI MCLKs
commit dc512ed34b67fdf4c4dd2efd31a56a15210ff404 from
https://github.com/nxp-imx/linux-imx

Register the SAI MCLKs based on what's found in their devicetree
node. If there is a board that needs to pass on some external clock
on one of the SAIx_MCLK pins, all they have to do is to override the
devicetree node for that specific saiX_mclk and the driver will pick
it up and be able to set it as parent inside the SAIs clocks hierarchy.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Reviewed-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>