dmaengine: fsl-dpaa2-qdma: Fix the size of dma pools
commit 3e859726b22c6a129682e37f444ae18db31a63d8 from
https://github.com/nxp-imx/linux-imx

In case of long format of qDMA command descriptor, there are one frame
descriptor, three entries in the frame list and two data entries. So the
size of dma_pool_create for these three fields should be the same with
the total size of entries respectively, or the contents may be overwritten
by the next allocated descriptor.

Signed-off-by: Guanhua Gao <guanhua.gao@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>