LF-3319 net: stmmac: enable rx_clk when killing the vlan
commit  2816b3b1ce51135f60ff8fa09eed93b1d04d2129 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

when killing the vlan, the rx_clk from the phy need to be enabled.
The patch checks the phy's status before killing the vlan. If the phy
is stopped, enable it, remove the vlan, and then stop the phy.

Reviewed-by: Frank Li <frank.li@nxp.com>
Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>