LF-7709-1 arm64: dts: imx8ulp: use a reserved mem-ranges to constrain ele-mu dma-range
commit  d82e2ac66622ccfce9081adca91a4192bc5675c3 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

EdgeLock Enclave are has a hardware limitation of restricted access
to the DDR memory range:
- 0x80000000  0x9FFFFFFF

ELE-MU driver requireis 1MB of memory. In this patch the we are reserving
1MB of ddr memory region from the lower 32-bit range.

Initially the driver was dependent on the dma-ranges property, added as
part of the commit: e24cdd398372.

Although dma-ranges property can restrict the dma-mask, but allocation
is not reliable. This due to the memory allocation strategy as the
followings:

-  dma-allocate_coherent
      -> alloc pages
          -> check dma regions
              -> re-alloc pages from DMA/32_ZONES->check dma regions.

The flow is alloc pages then check the regions, which will result in
returning failure when region checks failed in the last step. So, using
dma-regions is likely returned failure, and not reliable even if there
is plenty of memory in the system which is suitable.

This could cause the HSM operation to fail, resulting in a kernel crash.

This patch fixes the limitation associated with the dma-ranges property.

Since, original approach of using dma-ranges was not stable, this patch
should be squashed with the old commit.

Signed-off-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Pankaj Gupta <pankaj.gupta@nxp.com>
Reviewed-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>