dpaa2-switch: offload port objects on the BOND interface
commit  c33ab2f590b309d33cf949d981811ee1c44dc261 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

We have to intercept and act on events such as
SWITCHDEV_OBJ_ID_PORT_VLAN sent on a bond device which is the upper of
one of our switch ports.
When this happens, we have to add that specific VLAN requested to be
added on the bond device on all the DPAA2 switch interfaces which are
under that bond. We do this by calling switchdev_handle_port_obj_add
with the appropriate parameters.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>