dpaa2-switch: say that the bond device is offloaded as a bridge port
commit  d2b0b48b337eecc55f26217c2a7215bc84d99376 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

We have to let the switchdev framework know if a specific bridge port is
offloaded or not, even if that port is bond device.
For this to happen, create the dpaa2_switch_port_to_bridge_port function
which will determine the bridge port corresponding to a particulat dpaa2
switch interface and use it in the switchdev_bridge_port_offload call.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>