arm64: dts: ls1028a-rdb-dpdk: remove pause property
commit d68971e331fc564165e387e75cf58b19882de917 from
https://github.com/nxp-imx/linux-imx

DPDK is not capable to handle any type of congestion
or back-pressure. So if there is congestion on the network,
HW can generate PAUSE frames to limit the traffic and packets
can be dropped in the DPDK.

So in this patch, removing the pause property from switch and
ethernet ports.

Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>