LF-6841: dma: pxp: fix pxp timeout issue when run g2d samples test
commit 8dbfa6f08de4c0d22290fd1ed5112259c9b8cb74 from
https://github.com/nxp-imx/linux-imx.git

When run g2d sample test, pxp will report timeout issue. The reason
is that BGRA32 format is used in the case and driver will select AS
engine as input source. But from PXP datasheet, As can't do a simple
copy operation since it usually need to co-work with PS engine which
mean the output pixel value will determined by both AS and PS, such
as raster operation "AS OR PS" which usually used in legacy blending
operation. So add this workaround to fix this issue.

Reproduce steps:
/opt/g2d_samples/g2d_wayland_shm_test

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>