LF-8465 arm64: dts: imx8dxl-evk: fix the compatible of scu-gpio
commit 7fa68ad95d9684589a141b7593fdbf4abf542f32 from
https://github.com/nxp-imx/linux-imx.git

The compatible string for the scu-gpio driver was recently changed in
the upstreaming kernel to comply with standard naming rules. However,
this change was missed in the local repo, causing compatibility issues
on the imx8dxl-evk board.

To fix this issue, the compatible string was updated to "fsl,imx8qxp-sc-gpio"
in the device tree node for scu-gpio0.

Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>
Reviewed-by: Frank Li <frank.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>