LF-8907-1: i2c: imx-lpi2c: extend timeout time for the master to receive data
commit 715aac5554575763023d24c98ab875f8b184d87a from
https://github.com/nxp-imx/linux-imx.git

The master of LPI2C needs to read data from the slave by writing
the number of received data to txfifo. However, the TXFIFO register
only has one byte length, and if the length of the data that needs
to be read exceeds 256 bytes, it needs to be written to TXFIFO
multiple times. TXFIFO has a depth limit, a wait loop is needed to
prevent TXFIFO from overflowing.

The previous logic limited the receiving time of the entire message
to finish in 1 second. If the data is too long, and if there is the
transmission delay caused by CPUIDLE/high CPU loading, the data
transmission will fail. So reset the timeout timer after receiving
a set of 256bytes data each time.

Signed-off-by: Carlos Song <carlos.song@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>