[armel/marvell] Disable features to shrink kernel image (fixes FTBFS)
This is following the idea which Ben Hutchings applied earlier in
9dfe6d33a4fd ("[armel/marvell: Disable features to shrink kernel image
(fixes FTBFS)").

In the latest armel build for bookworm-security (6.1.52-1), the
compressed image for marvell increased and was about 24K too large:

   Image size 2753652/2729712, using 100.88%.  Too large.  Refusing to continue.

In the upstream changes Ben did found there were two commits in 6.2
which each increased the image size substantially:

- 60443c88f3a8 "kallsyms: Improve the performance of
  kallsyms_lookup_name()". Disabling KALLSYMS (and FUNCTION_TRACER
  which selects it) would save ~360K (!), but it looks like many user-
  space programs need /proc/kallsyms.  We already disable
  KALLSYMS_ALL to reduce the number of symbols included.
- f4d6b94b40c9 "apparmor: use zstd compression for profile data" made
  SECURITY_APPARMOR_EXPORT_BINARY select ZSTD_COMPRESS, which would
  otherwise be modular.  This option is mostly useful for checkpoint/
  restore which we already disable, so seems safe to disable.

For first of those was later on backported to 6.1.42. Following the same
reasoning, it might be possible to disable two of the features without
much impact:

- MPTCP costs ~40K and is mostly useful for mobile devices or servers
  supporting them.
- FUNCTION_TRACER costs ~90K and is mostly useful for debugging.

After disabling those two features

   Image size 2644124/2729712, using 96.86%.  Image fits.  Continuing.