net: fec: remove useless fec_enet_reset_skb()
commit b4455aa991488ca572d3756e8be1582c78ecbc4e from
https://github.com/nxp-imx/linux-imx.git

This patch is a cleanup for fec driver. The fec_enet_reset_skb()
is used to free skb buffers for tx queues and is only invoked in
fec_restart(). However, fec_enet_bd_init() also resets skb buffers
and is invoked in fec_restart() too. So fec_enet_reset_skb() is
redundant and useless.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Simon Horman <simon.horman@corigine.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>