net: phy: mtip_backplane: add driver for MoreThanIP backplane AN/LT core
commit  19aa941dc1bfcb28c8a41a9af3b888be30436000 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

For each networking SerDes lane on certain Layerscape SoCs, there is a
block, based on an IP core from MoreThanIP, which optionally handles
IEEE 802.3 clause 73 and clause 72, i.e. backplane auto-negotiation and
link training.

The hardware integration between the SerDes lane and this AN/LT block is
rather weak. For this reason, there is no automatic link training
performed in hardware, but rather, software needs to implement a custom,
SerDes-specific link training algorithm and use the AN/LT registers to
communicate it with the link partner. This driver is an inapt attempt to
do just that.

Since the MTIP AN/LT block may be, in premise, integrated in non-NXP
SoCs as well, the networking PHY driver implementation is as generic as
possible.

Initial support is present only for the LX2160A SoC and the single-lane
link modes (10G, 25G). For this SoC, the register map of the IP block
was a bit mangled, and we don't have any PHY ID. I didn't want to invent
one, so this driver probes on a specific compatible string, plus the
secondary "ethernet-phy-ieee802.3-c45" which makes the PHY library
recognize it as a clause 45 PHY and not an mdio_device.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>