LF-7770: remoteproc: imx_dsp_rproc: add mandatory find_loaded_rsc_table op
commit  63bd8fa873a2cdf56b202e78e54676d2f0f0a20b from
https://github.com/nxp-imx/linux-imx lf-6.1.y

Add the .find_loaded_rsc_table operation for i.MX DSP.
We need it for inter-process communication between DSP
and main core.

This callback is used to find the resource table (defined
in remote processor linker script) where the address of the
vrings along with the other allocated resources (carveouts etc)
are stored.
If this is not found, the vrings are not allocated and
the IPC between cores will not work.

Signed-off-by: Iuliana Prodan <iuliana.prodan@nxp.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>