phy: lynx-28g: truly power the lanes up or down
commit  2628c361d7fd904a82b9bf5d9fe5d2445949ed4f from
https://github.com/nxp-imx/linux-imx lf-6.1.y

The current procedure for power_off() and power_on() is the same as the
one used for major lane reconfiguration, aka halting. But one can
observe that a halted lane does not cause, for example, the CDR of the
link partner to lose lock.

Implement the procedure mentioned in the block guide for powering down
a lane, and then back on.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>