net: mdio: add definition for Parallel Detection Failure bit
commit  f177d41bcd6a8fe265e7be8f8f976bcfd90612f1 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

This is a bit in the C45 standard MDIO_STAT1 register which says that
Parallel Detection failed. Parallel Detection is the process by which a
link falls back in case there is no autoneg base page seen from the link
partner. The data stream is fed to all the available technology-dependent
PCSes, and if one of them sees a link, the link becomes up. The failure
is when more than one PCS detects a link based on the same RX stream.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>