phy: ethernet: add configuration interface for copper backplane Ethernet PHYs
commit  ae7386d1ebb9174f2a834262f7d06ea2516514d9 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

In Layerscape and QorIQ SoCs, compliance with the backplane Ethernet
protocol is bolted on top of the SerDes lanes using an external IP core,
that is modeled as a phylink_pcs. This means that dynamic tuning of the
electrical equalization parameters of the link needs to be communicated
with the consumer of the generic PHY.

Create a small layer of glue API between a phylink_pcs (dealing with the
AN/LT logic for backplanes) and a generic PHY by extending the
phy_configure() API with a new struct phy_configure_opts_ethernet.

There are 2 directions of interest. In the "local TX training", the
generic PHY consumer gets requests over the wire from the link partner
regarding changes we should make to our TX equalization. In the "remote
TX training" direction, the generic PHY is the producer of requests,
based on its RX status, and the generic PHY consumer polls for these
requests until we are happy. Each request is also sent (externally to
the generic PHY layer) to the link partner board, for it to adjust its
TX equalization.

struct phy_configure_opts_ethernet is valid when phy_set_mode_ext() has
been called with PHY_MODE_ETHERNET or PHY_MODE_ETHTOOL, same as with
other union phy_configure_opts types.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>