LF-9849-4: net: phy: tja11xx: call resume before soft reset.
commit 6d635d1095ab21d376952e1cab661c28d6eb2600 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git

The newly added condition check in the __phy_resume function is
causing failures when bringing the interface from the "down" state
to the "up" state.

+       if (!phydrv || !phydrv->resume || !phydev->suspended)
                                           ^^^^^^^^^^^^^^^^^
The genphy_soft_reset function will reset the phydev->suspended member
variable to 0. This would break the subsequent resume logic and put
the phy into a permanent powered down state.

To avoid this issue, the resume function should be called before the
soft reset. Doing so will synchronize the software states with the
hardware registers. This prevents the phy from getting stuck in a
powered down state.

The BMCR_PDOWN bit is set when initializing the tja1100 configuration
to avoid conflicts if both REFCLK_IN and REFCLK_OUT are enabled. In this
way, it can avoid the issue of the configuration unable to take effect
even after the registers are correctly programmed.

Fixes: f10880f9fa61 ("net: phy: balance calls to ->suspend() and ->resume()")
Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>
Reviewed-by: Frank Li <frank.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>