phy: lynx-28g: add algorithm for IEEE 802.3 C72 (10GBase-KR) link training
commit a72d50fd112e2d4661233b2fb0227af889cce3b2 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

Provide an implementation for the new phy_configure_opts_xgkr API for
the 28G Lynx SerDes from NXP LX2160A. The core logic of the algorithm is
also usable with the 10G Lynx SerDes for earlier QorIQ/Layerscape SoCs,
but a driver does not exist for that yet. Nonetheless, it is in progress,
so structure the link training algorithm as library code, shareable by
the two SerDes drivers when the time comes.

Loosely based on previous work by Florinel Iordache.

Co-developed-by: Florinel Iordache <florinel.iordache@nxp.com>
Signed-off-by: Florinel Iordache <florinel.iordache@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>