arch: arm64: lx2160a-rdb: comment out tx-disable-gpios from the SFP cage description
commit  6b4dc6d20b67143a55c3b7c34fbc88ef60981876 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

Comment out the tx-disable-gpios from the two SFP cages found on the
LX2160ARDB board. This is needed the first thing that the sfp subsystem
does when probing the cages is to disable the Tx signal using the gpio
described using the 'tx-disable-gpios' property.

Because in DPAA2's case the ports could be handled by the MC firmware
in case of TYPE_FIXED, no phylink_start() will be called, thus the Tx in
the SFP will remain disabled rendering a link down on the port.

In this case, just remove the gpio so that irrespective of the mode in
which the port runs (TYPE_FIXED, TYPE_PHY etc) a link up will get
established.

Fixes: c4c6978ac1c3 ("arm64: dts: lx2160a-rdb: describe the two SFP cages")
Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>