net/sched: taprio: pass mqprio queue configuration to ndo_setup_tc()
commit 5c2f9eff89b28d20211519bdbdd9800b125ab785 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git

The taprio offload does not currently pass the mqprio queue
configuration down to the offloading device driver. So the driver cannot
act upon the TXQ counts/offsets per TC, or upon the prio->tc map. It was
probably assumed that the driver only wants to offload num_tc (see
TC_MQPRIO_HW_OFFLOAD_TCS), which it can get from netdev_get_num_tc(),
but there's clearly more to the mqprio configuration than that.

To remedy that, we need to actually reconstruct a struct
tc_mqprio_qopt_offload to pass as part of the tc_taprio_qopt_offload.
The problem is that taprio doesn't keep a persistent reference to the
mqprio queue structure in its own struct taprio_sched, instead it just
applies the contents of that to the netdev state (prio:tc map, per-TC
TXQ counts and offsets, num_tc etc). Maybe it's easier to understand
why, when we look at the size of struct tc_mqprio_qopt_offload: 352
bytes on arm64. Keeping such a large structure would throw off the
memory accesses in struct taprio_sched no matter where we put it. So we
prefer to dynamically reconstruct the mqprio offload structure based on
netdev information, rather than saving a copy of it.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Alexandre Howard <alexandre.howard@nxp.com>
Link: https://lore.kernel.org/netdev/20230126125308.1199404-13-vladimir.oltean@nxp.com/
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>