phy: lynx-28g: truly power the lanes up or down
commit f73df580c2379a90c9269dace6dc3084fdf85ccc from
https://github.com/nxp-imx/linux-imx lf-6.1.y

The current procedure for power_off() and power_on() is the same as the
one used for major lane reconfiguration, aka halting. But one can
observe that a halted lane does not cause, for example, the CDR of the
link partner to lose lock.

Implement the procedure mentioned in the block guide for powering down
a lane, and then back on.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
[Xulin: add functions lynx_28g_lane_halt() & lynx_28g_lane_reset()
directly, since they were mistakenly removed by the previous merge.]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>