debian/salsa-ci.yml: Reuse the common provisioning codes in the pipeline/salsa-ci.yml
We have merged the kernel source code into the Debian kernel git repo,
so it is possible to reuse the common provisioning codes in the
pipeline/salsa-ci.yml. We only need to add a 'after_script' to generate
the debian/control. With this change, the source unpack is also not
needed any more in build stage.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>