debian/salsa-ci.yml: Increase the number of concurrent jobs for compiling the kernel
It may takes a very long time to build the kernel, especially for the
non-x86 architecture, for example, more than 3 hours even for a
pkg.kernel.quick profile build on arm64. So increate the number of
concurrent jobs for compiling the kernel. At the same time, also
increase the timeout time for the build job.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>