debian/salsa-ci.yml: Don't need to unpack the kernel source in kconfig-static
With the changes introduced by commit 513808a19a52
("debian/salsa-ci.yml: Reuse the common provisioning codes in the
pipeline/salsa-ci.yml"), we choose to pass the unpacked kernel source
to the next stage in extract-source, so there is no need to do the
kernel unpack in other jobs. All we need to do is just using
"dpkg-source --before-build" to make sure all the kernel patches
are applied.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>