debian/changelog: Add wrX suffix for the upstream version
We have merged the kernel source into the Debian kernel git repo, that
also means that the current kernel git repo is a git packaging
repository. So for each time we tweak the kernel source codes, there
will be a new upstream version. For this reason we can't use the vanilla
kernel version as our upstream kernel version anymore. Add a wrX suffix
to the current upstream version and we should bump this suffix version
for each time the kernel source is changed.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>