debian: Add imx infix for the kbuild package
The kbuild package is essential for the building of signed package,
so add a imx infix for this package name to make sure there is no
name conflict between two different BSPs. This only change the package
name, we choose to keep the original directories of installed files,
on purpose, otherwise it may break the tools such as kernel-team.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>