debian/salsa-ci.yml: Disable publish to shared aptly for base branch
The base branch is intended to stage common changes which can be
applied to all other branches, such as amd64 or nxp. However, we don't
want to populate any packages build from this branch to shared aptly.
Therefore, updating the debian/changelog on this branch is unnecessary,
The shared-aptly job would fail due to the unchanged kernel version,
and more critically, this failure will disrupt the upload of all packages
to shared aptly. To address this, disable publish to shared aptly for
base branch.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>